requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.1
redis[hiredis]==5.0.1
cryptography==41.0.5
pyjwt==2.8.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
ollama==0.1.7
gitpython==3.1.40
prometheus-client==0.19.0
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
